requirements.txt
Pillow==11.1.0
pyserial==3.5
